Refactored libConCoct into a package with multiple modules.
