+++ libConCoct/__init__.py
@@ -1,0 +1,5 @@
+"""libConCoct - Builds simple C programs and runs unit tests
+
+.. codeauthor:: Martin Wichmann
+.. codeauthor:: Christian Wichmann
+"""
